Moved JSON serialization logic into the EventSerializer for better testability.
