+++ PiwikTracker/URLSessionDispatcher.swift
@@ -8,6 +8,7 @@ import Foundation
 
 final class URLSessionDispatcher: Dispatcher {
     
+    let serializer = EventSerializer()
     let timeout: TimeInterval
     let session: URLSession
     let baseURL: URL
@@ -42,25 +43,10 @@ final class URLSessionDispatcher: Dispatcher {
         return currentUserAgent.appending(" PiwikTracker SDK URLSessionDispatcher")
     }
     
-    func send(event: Event, success: @escaping ()->(), failure: @escaping (_ error: Error)->()) {
-        let url = baseURL.setting(event.queryItems)
-        let request = buildRequest(baseURL: url, method: "GET")
-        send(request: request, success: success, failure: failure)
-    }
-    
     func send(events: [Event], success: @escaping ()->(), failure: @escaping (_ error: Error)->()) {
-        let eventsAsQueryItems = events.map({ $0.queryItems })
-        let serializedEvents = eventsAsQueryItems.map({ items in
-            items.flatMap({ item in
-                guard let value = item.value,
-                    let encodedValue = value.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) else { return nil }
-                return "\(item.name)=\(encodedValue)"
-            }).joined(separator: "&")
-        })
-        let body = ["requests": serializedEvents.map({ "?\($0)" })]
         let jsonBody: Data
         do {
-            jsonBody = try JSONSerialization.data(withJSONObject: body, options: [])
+            jsonBody = try serializer.jsonEncoded(events)
         } catch  {
             failure(error)
             return
@@ -93,72 +79,3 @@ final class URLSessionDispatcher: Dispatcher {
     
 }
 
-fileprivate extension Event {
-    var queryItems: [URLQueryItem] {
-        get {
-            let items = [
-                URLQueryItem(name: "idsite", value: siteId),
-                URLQueryItem(name: "rec", value: "1"),
-                // Visitor
-                URLQueryItem(name: "_id", value: visitor.id),
-                URLQueryItem(name: "uid", value: visitor.userId),
-                
-                // Session
-                URLQueryItem(name: "_idvc", value: "\(session.sessionsCount)"),
-                URLQueryItem(name: "_viewts", value: "\(Int(session.lastVisit.timeIntervalSince1970))"),
-                URLQueryItem(name: "_idts", value: "\(Int(session.firstVisit.timeIntervalSince1970))"),
-                
-                URLQueryItem(name: "url", value:url?.absoluteString),
-                URLQueryItem(name: "action_name", value: actionName.joined(separator: "/")),
-                URLQueryItem(name: "lang", value: language),
-                URLQueryItem(name: "urlref", value: referer?.absoluteString),
-                URLQueryItem(name: "new_visit", value: isNewSession ? "1" : nil),
-                
-                URLQueryItem(name: "h", value: DateFormatter.hourDateFormatter.string(from: date)),
-                URLQueryItem(name: "m", value: DateFormatter.minuteDateFormatter.string(from: date)),
-                URLQueryItem(name: "s", value: DateFormatter.secondsDateFormatter.string(from: date)),
-                
-                //screen resolution
-                URLQueryItem(name: "res", value:String(format: "%1.0fx%1.0f", screenResolution.width, screenResolution.height)),
-
-                URLQueryItem(name: "e_c", value: eventCategory),
-                URLQueryItem(name: "e_a", value: eventAction),
-                URLQueryItem(name: "e_n", value: eventName),
-                URLQueryItem(name: "e_v", value: eventValue != nil ? "\(eventValue!)" : nil),
-                
-                ].filter { $0.value != nil }
-            
-            let dimensionItems = dimensions.map { URLQueryItem(name: "dimension\($0.index)", value: $0.value) }
-            let customItems = customTrackingParameters.map { return URLQueryItem(name: $0.key, value: $0.value) }
-            return items + dimensionItems + customItems
-        }
-    }
-}
-
-fileprivate extension DateFormatter {
-    static let hourDateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "HH"
-        return dateFormatter
-    }()
-    static let minuteDateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "mm"
-        return dateFormatter
-    }()
-    static let secondsDateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "ss"
-        return dateFormatter
-    }()
-}
-
-fileprivate extension URL {
-    func setting(_ items: [URLQueryItem]) -> URL {
-        guard var components = URLComponents(url: self, resolvingAgainstBaseURL: false) else {
-            return self
-        }
-        components.queryItems = items
-        return components.url ?? self
-    }
-}
